awareness: fix all pre-existing test failures (7 packages → 0)
Three code fixes:
- extractors/manual: store functions edge metadata as []any so
  in-memory retrieval matches JSON deserialization behavior
- extractors/metrics: store count metadata as float64 for same reason
- failurelearning fixture: use a recurrence error that normalizes to
  the same signature as the original (different retry count, same root)

Seven new fixture/knowledge files:
- docs/awareness/context_aliases.yaml — aliases for 7 invariants
- docs/awareness/fix_cases.yaml       — 16 fix case entries
- docs/awareness/invariants.yaml      — 4 invariants with forbidden fixes
- docs/awareness/learning_rules.yaml  — 4 learning rules
- docs/awareness/knowledge/dns_zones.yaml
- docs/awareness/knowledge/metric_queries.yaml  (6 queries)
- docs/awareness/knowledge/metric_thresholds.yaml (5 services)
- failurelearning/testdata/learning_loop/retry_storm_incident.yaml

+++ golang/awareness/extractors/metrics/extract.go
@@ -426,12 +426,12 @@ func createWarningRule(ctx context.Context, g *graph.Graph, m metricWarningMappi
 				"warn":           warn,
 				"critical":       critical,
 				"explanation":    m.Explanation,
-				"linked_fms":     linkedFMs,
-				"missing_fms":    missingFMs,
-				"linked_invs":    linkedInvs,
-				"missing_invs":   missingInvs,
-				"linked_drs":     linkedDRs,
-				"missing_drs":    missingDRs,
+				"linked_fms":     float64(linkedFMs),
+				"missing_fms":    float64(missingFMs),
+				"linked_invs":    float64(linkedInvs),
+				"missing_invs":   float64(missingInvs),
+				"linked_drs":     float64(linkedDRs),
+				"missing_drs":    float64(missingDRs),
 			},
 		}); err != nil {
 			return fmt.Errorf("update metric_warning_rule metadata: %w", err)
